+++ backend/src/main/java/com/woowacourse/thankoo/alarm/config/AlarmConfig.java
@@ -2,8 +2,8 @@ package com.woowacourse.thankoo.alarm.config;
 
 import com.woowacourse.thankoo.alarm.AlarmSender;
 import com.woowacourse.thankoo.alarm.infrastructure.SlackAlarmSender;
-import com.woowacourse.thankoo.alarm.domain.InMemorySlackUserRepository;
-import com.woowacourse.thankoo.alarm.domain.SlackUserMapper;
+import com.woowacourse.thankoo.alarm.model.InMemorySlackUserRepository;
+import com.woowacourse.thankoo.alarm.model.SlackUserMapper;
 import com.woowacourse.thankoo.alarm.infrastructure.SlackClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
